+++ src/blob/index/bptree/core.rs
@@ -23,7 +23,10 @@ pub(crate) struct BPTreeFileIndex<K> {
 }
 
 #[async_trait::async_trait]
-impl<K: Key + 'static> FileIndexTrait<K> for BPTreeFileIndex<K> {
+impl<K> FileIndexTrait<K> for BPTreeFileIndex<K>
+where
+    for<'a> K: Key<'a> + 'static,
+{
     async fn from_file(name: FileName, ioring: Option<Rio>) -> Result<Self> {
         trace!("open index file");
         let file = File::open(name.to_path(), ioring)
@@ -48,9 +51,10 @@ impl<K: Key + 'static> FileIndexTrait<K> for BPTreeFileIndex<K> {
         headers: &InMemoryIndex<K>,
         meta: Vec<u8>,
         recreate_index_file: bool,
+        blob_size: u64,
     ) -> Result<Self> {
         clean_file(path, recreate_index_file)?;
-        let res = Self::serialize(headers, meta)?;
+        let res = Self::serialize(headers, meta, blob_size)?;
         let (mut header, metadata, buf) = res;
         let file = File::create(path, ioring)
             .await
@@ -107,9 +111,9 @@ impl<K: Key + 'static> FileIndexTrait<K> for BPTreeFileIndex<K> {
         self.read_headers(leaf_offset, key, &mut buf).await
     }
 
-    async fn get_records_headers(&self) -> Result<(InMemoryIndex<K>, usize)> {
+    async fn get_records_headers(&self, blob_size: u64) -> Result<(InMemoryIndex<K>, usize)> {
         let mut buf = self.file.read_all().await?;
-        self.validate_header(&mut buf).await?;
+        self.validate_header(&mut buf, blob_size).await?;
         let offset = self.metadata.leaves_offset as usize;
         let records_end = FileIndexTrait::<K>::file_size(self) as usize;
         let records_buf = &buf[offset..records_end];
@@ -137,21 +141,42 @@ impl<K: Key + 'static> FileIndexTrait<K> for BPTreeFileIndex<K> {
         self.read_header(leaf_offset, key, &mut buf).await
     }
 
-    fn validate(&self) -> Result<()> {
+    fn validate(&self, blob_size: u64) -> Result<()> {
         // FIXME: check hash here?
         if !self.header.is_written() {
             let param = ValidationErrorKind::IndexIsWritten;
-            return Err(Error::validation(param, "Index Header version is not valid").into());
+            return Err(
+                Error::validation(param, "Index is incomplete (no 'is_written' flag)").into(),
+            );
         }
         if self.header.version() != HEADER_VERSION {
             let param = ValidationErrorKind::IndexVersion;
             return Err(Error::validation(param, "Index Header version is not valid").into());
         }
+        if self.header.blob_size() != blob_size {
+            let param = ValidationErrorKind::IndexBlobSize;
+            return Err(Error::validation(
+                param,
+                format!(
+                    "Index Header is for blob of size {}, but actual blob size is {}",
+                    self.header.blob_size(),
+                    blob_size
+                ),
+            )
+            .into());
+        }
+        if self.header.magic_byte() != INDEX_HEADER_MAGIC_BYTE {
+            let param = ValidationErrorKind::IndexMagicByte;
+            return Err(Error::validation(param, "Index magic byte is not valid").into());
+        }
         Ok(())
     }
 }
 
-impl<K: Key + 'static> BPTreeFileIndex<K> {
+impl<K> BPTreeFileIndex<K>
+where
+    for<'a> K: Key<'a> + 'static,
+{
     async fn find_leaf_node(&self, key: &K, mut offset: u64, buf: &mut [u8]) -> Result<u64> {
         while offset < self.metadata.leaves_offset {
             offset = if offset == self.metadata.tree_offset {
@@ -230,8 +255,8 @@ impl<K: Key + 'static> BPTreeFileIndex<K> {
             let m_off = record_header_size * m as usize;
             let record_end = m_off + record_header_size;
             let record_header: RecordHeader = deserialize(&raw_headers_buf[m_off..record_end])?;
-            let read_key = record_header.key().to_vec().into();
-            match key.cmp(&read_key) {
+            let cmp_res = key.as_ref_key().cmp(&record_header.key().into());
+            match cmp_res {
                 CmpOrdering::Less => r = m - 1,
                 CmpOrdering::Greater => l = m + 1,
                 CmpOrdering::Equal => return Ok(Some((record_header, m_off))),
@@ -334,8 +359,8 @@ impl<K: Key + 'static> BPTreeFileIndex<K> {
         Ok(())
     }
 
-    async fn validate_header(&self, buf: &mut Vec<u8>) -> Result<()> {
-        self.validate()?;
+    async fn validate_header(&self, buf: &mut Vec<u8>, blob_size: u64) -> Result<()> {
+        self.validate(blob_size)?;
         if !Self::hash_valid(&self.header, buf)? {
             let param = ValidationErrorKind::IndexChecksum;
             return Err(Error::validation(param, "header hash mismatch").into());
@@ -380,9 +405,10 @@ impl<K: Key + 'static> BPTreeFileIndex<K> {
     fn serialize(
         headers_btree: &InMemoryIndex<K>,
         meta: Vec<u8>,
+        blob_size: u64,
     ) -> Result<(IndexHeader, TreeMeta, Vec<u8>)> {
         Serializer::new(headers_btree)
-            .header_stage(meta)?
+            .header_stage(meta, blob_size)?
             .tree_stage()?
             .build()
     }