+++ src/record.rs
@@ -1,6 +1,7 @@
 use crate::{error::ValidationErrorKind, prelude::*};
 
 pub(crate) const RECORD_MAGIC_BYTE: u64 = 0xacdc_bcde;
+const DELETE_FLAG: u8 = 0x01;
 
 #[derive(Serialize, Deserialize, Default, Clone, PartialEq)]
 pub struct Record {
@@ -89,7 +90,10 @@ impl Record {
     }
 
     /// Creates new `Record` with provided data, key and meta.
-    pub fn create<K: Key>(key: &K, data: Vec<u8>, meta: Meta) -> bincode::Result<Self> {
+    pub fn create<K>(key: &K, data: Vec<u8>, meta: Meta) -> bincode::Result<Self>
+    where
+        for<'a> K: Key<'a>,
+    {
         let key = key.as_ref().to_vec();
         let meta_size = meta.serialized_size()?;
         let data_checksum = CRC32C.checksum(&data);
@@ -111,6 +115,15 @@ impl Record {
         buf.extend(&raw_meta);
         buf.extend(&self.data);
         Ok(buf)
+    }
+
+    pub(crate) fn deleted<K>(key: &K) -> bincode::Result<Self>
+    where
+        for<'a> K: Key<'a> + 'static,
+    {
+        let mut record = Record::create(key, vec![], Meta::default())?;
+        record.header.mark_as_deleted()?;
+        Ok(record)
     }
 
     pub(crate) fn set_offset(&mut self, offset: u64) -> bincode::Result<()> {
@@ -268,5 +281,18 @@ impl Header {
         self.check_header_checksum()
             .with_context(|| "check header checksum failed")?;
         Ok(())
+    }
+
+    pub(crate) fn mark_as_deleted(&mut self) -> bincode::Result<()> {
+        self.flags |= DELETE_FLAG;
+        self.update_checksum()
+    }
+
+    pub(crate) fn is_deleted(&self) -> bool {
+        self.flags & DELETE_FLAG == DELETE_FLAG
+    }
+
+    pub(crate) fn created(&self) -> u64 {
+        self.created
     }
 }