+++ CHANGELOG.md
@@ -6,15 +6,48 @@ Pearl changelog
 #### Added
 - Add hierarchical filters support (#126)
 - Add tools to recover and migrate blobs and indexes (#148)
-
+- Add blob size to index header (#153)
+- Add magic byte to index header (#152)
 
 #### Changed
 
-
 #### Fixed
-
+- Fix delete with no active blob set (#167)
 
 #### Updated
+
+
+## [0.11.0] - 2022-04-18
+#### Added
+Add support for hierarchical range filters (#154)
+
+#### Changed
+- Remove active checking for blob update (#159)
+
+#### Fixed
+- corrupted blob now should be saved in case of 'unexpected eof' error (#160)
+
+#### Updated
+
+
+## [0.10.0] - 2022-03-30
+#### Changed
+- Add refkey trait (#141)
+- Change nightly to stable rust toolchain
+
+
+## [0.9.2] - 2022-02-14
+#### Added
+- Add record delete api method (#103)
+
+
+## [0.9.1] - 2022-02-03
+#### Added
+- Add wait cycle for tests with index files (#144)
+
+
+#### Fixed
+- In memory index last record retrieval (#147)
 
 
 ## [0.9.0] - 2021-12-09